Move job models into model directory
