--- piezo_web_app/PiezoWebApp/src/services/spark_job/validation/validation_result.py
@@ -1,6 +1,0 @@
-class ValidationResult:
-
-    def __init__(self, is_valid, message, validated_value):
-        self.is_valid = is_valid
-        self.message = message
-        self.validated_value = validated_value
